+++ services/web/server/src/simcore_service_webserver/cli.py
@@ -24,7 +24,7 @@ from aiohttp import web
 from models_library.basic_types import BuildTargetEnum
 
 from .application import create_application, run_service
-from .application_config import CLI_DEFAULT_CONFIGFILE, app_schema
+from .application__schema import CLI_DEFAULT_CONFIGFILE, app_schema
 from .application_settings import ApplicationSettings
 from .cli_config import add_cli_options, config_from_options
 from .log import setup_logging
