+++ services/web/server/tests/unit/isolated/test_catalog_setup.py
@@ -7,7 +7,7 @@ import pytest
 from servicelib.aiohttp.application import create_safe_application
 from servicelib.aiohttp.client_session import APP_CLIENT_SESSION_KEY
 from simcore_service_webserver._meta import api_version_prefix
-from simcore_service_webserver.application_config import load_default_config
+from simcore_service_webserver.application__schema import load_default_config
 from simcore_service_webserver.catalog import (
     is_service_responsive,
     setup_catalog,
